sync library from upstream (ABI 176 + object-navigation work)
Full resync of the binding source from the upstream dev repo, which is the
authoritative source. Brings the ABI 176 layout/signature overhaul, the
object-navigation improvements (Function callers/callees fix, IL
instruction Function/Parent/Instr/Tokens, IL-at-address helpers,
BinaryView iteration + address-based xrefs, Variable/DataVariable/Type/
Symbol navigation) and the IL instruction-factory robustness fix (no
longer throws on unhandled operations).

Packaging files (BinaryNinja.csproj, README, LICENSE, .github,
.globalconfig, .gitignore) are preserved. Builds clean (0/0).

+++ Enumeration/BNActionType.cs
@@ -5,7 +5,7 @@ namespace BinaryNinja
 	/// <summary>
 	/// 
 	/// </summary>
-    public enum ActionType : uint
+    public enum ActionType : byte
 	{
 		/// <summary>
 		///
